C restructure
