+++ CutePy/src/cute_angular.c
@@ -1,63 +1,6 @@
-#include "cute_drivers.h"
+#include "config.h"
+#include "utils.h"
 
-int NodeThis=0;
-
-static void *my_malloc(size_t size)
-{
-  void *outptr=malloc(size);
-  if(outptr==NULL) {
-    fprintf(stderr,"CUTE: out of memory!\n");
-    exit(1);
-  }
-
-  return outptr;
-}
-
-static void *my_calloc(size_t nmemb,size_t size)
-{
-  void *outptr=calloc(nmemb,size);
-  if(outptr==NULL) {
-    fprintf(stderr,"CUTE: out of memory!\n");
-    exit(1);
-  }
-
-  return outptr;
-}
-
-static void print_info(char *fmt,...)
-{
-  if(NodeThis==0) {
-    va_list args;
-    char msg[256];
-    
-    va_start(args,fmt);
-    vsprintf(msg,fmt,args);
-    va_end(args);
-    
-    printf("%s",msg);
-  }
-}
-
-CuteBin *cute_bin_new(int nr, int is_log, double rmax, double rmin_log, double unit)
-{
-  CuteBin *bin=(CuteBin *)my_malloc(sizeof(CuteBin));
-  bin->nbins=nr;
-  bin->is_log=is_log;
-  bin->r_max=rmax*unit;
-  bin->i_r_max=1./bin->r_max;
-  bin->log_r_max=log10(bin->r_max);
-  if(is_log)
-    bin->n_logint=bin->nbins/log10(rmax/rmin_log);
-  else
-    bin->n_logint=0;
-  return bin;
-}
-
-void cute_bin_free(CuteBin *bin)
-{
-  free(bin);
-  return;
-}
 
 typedef struct {
   int nside_cth;
