+++ CutePy/src/utils.h
@@ -1,0 +1,12 @@
+#ifndef _CUTE_UTILS_
+#define _CUTE_UTILS_
+
+#include "cute.h"
+
+void *my_malloc(size_t size);
+
+void *my_calloc(size_t nmemb,size_t size);
+
+void print_info(char *fmt,...);
+
+#endif //_CUTE_UTILS_
